Update tempgifile.gi
Added a missing semicolon at the end of the ViewString method for a Monoid Character Table. Added Cases to the ViewSting of a Monoid Character. When a Monoid Character is the projective cover of another without knowing the values of the character of the cover then a lazy message is created.